[add] Nativeのビルド設定を追加

+++ buffer/cpu/cpp/src/cpu/collection_fun.h
@@ -1,13 +1,11 @@
 #ifndef COLLECTION_FUN_H
 #define COLLECTION_FUN_H
 
-#include <cstddef>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-float average_d1(const float* x, size_t size);
+float average_d1(const float* x, int size);
 
 void average_d2(const float* x, int xi, int xj, int axis, float* result);
 
@@ -15,7 +13,7 @@ void average_d3(const float* x, int xi, int xj, int xk, int axis, float* result)
 
 void average_d4(const float* x, int xi, int xj, int xk, int xl, int axis, float* result);
 
-float max_d1(const float* x, size_t size);
+float max_d1(const float* x, int size);
 
 void max_d2(const float* x, int xi, int xj, int axis, float* result);
 
@@ -23,7 +21,7 @@ void max_d3(const float* x, int xi, int xj, int xk, int axis, float* result);
 
 void max_d4(const float* x, int xi, int xj, int xk, int xl, int axis, float* result);
 
-float min_d1(const float* x, size_t size);
+float min_d1(const float* x, int size);
 
 void min_d2(const float* x, int xi, int xj, int axis, float* result);
 
@@ -31,7 +29,7 @@ void min_d3(const float* x, int xi, int xj, int xk, int axis, float* result);
 
 void min_d4(const float* x, int xi, int xj, int xk, int xl, int axis, float* result);
 
-float sum_d1(const float* x, size_t size);
+float sum_d1(const float* x, int size);
 
 void sum_d2(const float* x, int xi, int xj, int axis, float* result);
